requirements.txt
python-telegram-bot==22.7
reportlab==4.1.0
arabic-reshaper==3.0.0
python-bidi==0.4.2
gspread==5.12.4
google-auth==2.29.0
oauth2client==4.1.3